requirements.txt
streamlit
Pillow
numpy
onnxruntime
opencv-python-headless
transformers
torch
torchvision
scikit-learn
roboflow
